# float4 with better vectorization for encoder_forward.cu