--- a/dev/cuda/encoder_forward.cu
+++ b/dev/cuda/encoder_forward.cu
@@ -89,14 +89,6 @@ __device__ inline float4 operator+(const float4& a, const float4& b) {
     return make_float4(a.x + b.x, a.y + b.y, a.z + b.z, a.w + b.w);
 }
 
-__device__ inline float& vec_at(float4& vec, int index) {
-    return reinterpret_cast<float*>(&vec)[index];
-}
-
-__device__ inline float vec_at(const float4& vec, int index) {
-    return reinterpret_cast<const float*>(&vec)[index];
-}
-
 __global__ void encoder_forward_kernel3(float4* out,
                                const int* inp, const float4* wte, const float4* wpe,
                                int B, int T, int C) {
@@ -111,19 +103,10 @@ __global__ void encoder_forward_kernel3(float4* out,
         int c4 = idx % C4;
 
         int ix = inp[b * T + t];
-
-        // float4* out_btc4 = out + b * T * C4 + t * C4 + c4;
-        // const float4* wte_ix4 = wte + ix * C4 + c4;
-        // const float4* wpe_tc4 = wpe + t * C4 + c4;
-        // for (int i = 0; i < 4; i++) {
-        //     vec_at(*out_btc4, i) = vec_at(*wte_ix4, i) + vec_at(*wpe_tc4, i);
-        // }
-        //*out_btc4 = *wte_ix4 + *wpe_tc4;
+        
         out[b * T * C4 + t * C4 + c4] = wte[ix * C4 + c4] + wpe[t * C4 + c4];
     }
 }
-
-
 
 // ----------------------------------------------------------------------------
 // kernel launcher
@@ -154,7 +137,7 @@ void encoder_forward3(float* out,
                      const int block_size) {
     assert(C % 4 == 0);
     const int N = B * T * C;
-    const int grid_size = ceil_div(N, block_size);
+    const int grid_size = ceil_div(N / 4, block_size);
     encoder_forward_kernel3<<<grid_size, block_size>>>((float4*) out, inp, (float4*) wte, (float4*) wpe, B, T, C);
     cudaCheck(cudaGetLastError());
 }