On RTX 3070
Kernel 2
```
block_size   32 | time 0.2933 ms | bandwidth 343.26 GB/s
block_size   64 | time 0.2099 ms | bandwidth 479.50 GB/s
block_size  128 | time 0.1924 ms | bandwidth 523.24 GB/s
block_size  256 | time 0.1979 ms | bandwidth 508.77 GB/s
block_size  512 | time 0.1947 ms | bandwidth 516.96 GB/s
block_size 1024 | time 0.2276 ms | bandwidth 442.33 GB/s
```

Kernel 3 runtime is more stable across different blocksize and is faster in terms of the best execution time.
```
block_size   32 | time 0.1902 ms | bandwidth 529.21 GB/s
block_size   64 | time 0.1908 ms | bandwidth 527.52 GB/s
block_size  128 | time 0.1919 ms | bandwidth 524.55 GB/s
block_size  256 | time 0.1916 ms | bandwidth 525.45 GB/s
block_size  512 | time 0.1923 ms | bandwidth 523.43 GB/s
block_size 1024 | time 0.1896 ms | bandwidth 530.82 GB/s
```